+++ example/artist_methods_example.dart
@@ -52,7 +52,8 @@ void main() async {
   ));
 
   print(
-      'Artist Info Name: ${artistGetInfo.name} | Artist URL: ${artistGetInfo.url}');
+      'Artist Info Name: ${artistGetInfo.name} | Artist URL: ${artistGetInfo
+          .url}');
 
   print('########################artist.getSimilar###########################');