+++ lib/scrobblenaut.dart
@@ -4,6 +4,7 @@
  */
 
 /// Scrobblenaut Library.
+/// This is the brain of my Wrapper.
 library scrobblenaut;
 
 export 'package:scrobblenaut/src/scrobblenaut.dart';