+++ lib/scrobblenaut_helpers.dart
@@ -1,0 +1,11 @@
+/**
+ * Scrobblenaut - A deadly simple Last.FM API Wrapper for Dart.
+ * Copyright (c) 2020 Nebulino
+ */
+
+/// Scrobblenaut Library.
+/// Import this if you want to use some helpers.
+library scrobblenaut_helpers;
+
+export 'package:scrobblenaut/src/helpers/scrobble_response.dart';
+export 'package:scrobblenaut/src/helpers/scrobbled_track.dart';