+++ lib/src/lastfm/artist.dart
@@ -84,4 +84,6 @@ class Artist {
   });
 
   factory Artist.fromJson(Map<String, dynamic> json) => _$ArtistFromJson(json);
+
+  Map<String, dynamic> toJson() => _$ArtistToJson(this);
 }